ApiKeyGenPrefix: rename folder to aling with package name
